move DeveloperProject from CodeGenerator to Core
