+++ plugins/de.cognicrypt.core/src/de/cognicrypt/utils/DeveloperProject.java
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
 
-package de.cognicrypt.codegenerator;
+package de.cognicrypt.utils;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
@@ -26,6 +26,7 @@ import org.eclipse.jdt.core.IPackageFragmentRoot;
 import org.eclipse.jdt.core.JavaCore;
 
 import de.cognicrypt.core.Constants;
+import de.cognicrypt.core.Activator;
 
 /**
  * This class represents the app developer's project, on which the plugin is working.
@@ -97,7 +98,7 @@ public class DeveloperProject {
 	 *        name of the package
 	 * @return package as {@link org.eclipse.jdt.core.IPackageFragment IPackageFragment}
 	 * @throws CoreException
-	 *         see {@link de.cognicrypt.codegenerator.DeveloperProject#getSourcePath() getSourcePath()}
+	 *         see {@link de.cognicrypt.utils.DeveloperProject#getSourcePath() getSourcePath()}
 	 */
 	public IPackageFragment getPackagesOfProject(final String name) throws CoreException {
 		return JavaCore.create(this.project).getPackageFragmentRoot(this.project.getFolder(getSourcePath())).getPackageFragment(name);
